user_service/requirements.txt
fastapi
sqlalchemy
uvicorn
passlib[bcrypt]
pyjwt
psycopg2-binary
sqladmin
pika
pydantic[email]
loguru
pytest-cov
alembic
